+++ scripts/ammunition-system/actions/reload.js
@@ -2,7 +2,7 @@ import { handleReload } from "../../feats/crossbow-feats.js";
 import { getControlledActorAndToken, getEffectFromActor, getFlag, getItem, postInChat, setEffectTarget, showWarning, Updates, useAdvancedAmmunitionSystem } from "../../utils/utils.js";
 import { getWeapon, getWeapons } from "../../utils/weapon-utils.js";
 import { CONJURED_ROUND_EFFECT_ID, LOADED_EFFECT_ID, MAGAZINE_LOADED_EFFECT_ID, RELOAD_AMMUNITION_IMG } from "../constants.js";
-import { buildLoadedEffectName, checkFullyLoaded, isFullyLoaded } from "../utils.js";
+import { buildLoadedEffectName, checkFullyLoaded, isFullyLoaded, updateAmmunitionQuantity } from "../utils.js";
 import { setLoadedChamber } from "./next-chamber.js";
 import { selectAmmunition } from "./switch-ammunition.js";
 import { unloadAmmunition } from "./unload.js";
@@ -202,7 +202,7 @@ async function performReload(actor, token, weapon, updates) {
             }
 
             // Remove one piece of ammunition from the stack
-            updates.update(ammo, { "system.quantity": ammo.quantity - 1 });
+            updateAmmunitionQuantity(updates, ammo, -1)
         }
     } else {
         if (checkFullyLoaded(actor, weapon)) {
@@ -272,7 +272,7 @@ async function getAmmunition(weapon, updates) {
             false,
             false
         );
-    } else if (ammunition.quantity < 1) {
+    } else if (ammunition.quantity < 1 && ammunition.autoDestroy) {
         return await selectAmmunition(
             weapon,
             updates,